Update NpgsqlDatabaseModelFactory.cs
I have included 'p' in the 'cls.relkind' condition to also retrieve base partition tables. This is particularly useful for creating dynamic tables at runtime, eliminating the need for scaffolding and building each time. I tested this with the base table to check if it was insertable and confirmed that it is. Subsequently, the records were distributed into the partitioned tables, which do not need to be scaffolded every time.

Here is the modified condition:
cls.relkind IN ('r', 'v', 'm', 'f', 'p') -- added 'p' here